Make one ErrorViewModel class
