+++ SunEngine/Controllers/ArticlesController.cs
@@ -15,12 +15,12 @@ namespace SunEngine.Controllers
     public class ArticlesController : BaseController
     {
         private readonly OperationKeysContainer OperationKeys;
-        
+
         private readonly ArticlesOptions articlesOptions;
         private readonly ICategoriesStore categoriesStore;
         private readonly IAuthorizationService authorizationService;
         private readonly CategoriesAuthorization categoriesAuthorization;
-        
+
         private readonly ArticlesPresenter articlesPresenter;
 
 
@@ -29,12 +29,12 @@ namespace SunEngine.Controllers
             IAuthorizationService authorizationService,
             ICategoriesStore categoriesStore,
             OperationKeysContainer operationKeysContainer,
-            ArticlesPresenter articlesPresenter, 
+            ArticlesPresenter articlesPresenter,
             MyUserManager userManager,
             IUserGroupStore userGroupStore) : base(userGroupStore, userManager)
         {
             this.OperationKeys = operationKeysContainer;
-            
+
             this.articlesOptions = articlesOptions.Value;
             this.authorizationService = authorizationService;
             this.categoriesStore = categoriesStore;
@@ -56,20 +56,10 @@ namespace SunEngine.Controllers
                 return Unauthorized();
             }
 
-            IPagedList<ArticleInfoViewModel> articles = await articlesPresenter.GetArticlesAsync(category.Id,page,articlesOptions.ArticlesCategoryPageSize);
+            IPagedList<ArticleInfoViewModel> articles =
+                await articlesPresenter.GetArticlesAsync(category.Id, page, articlesOptions.ArticlesCategoryPageSize);
 
             return Json(articles);
         }
-    }
-    
-    public class ArticleInfoViewModel
-    {
-        public int Id { get; set; }
-        public string Title { get; set; }
-        public string AuthorName { get; set; }
-        public int MessagesCount { get; set; }
-        public DateTime PublishDate { get; set; }
-        public string CategoryTitle { get; set; }
-        public string CategoryName { get; set; }
     }
 }
